refactor(api): restructure auth module into proper modular monolith layout
Reorganises the flat src/modules/auth/*.ts pile into the correct subdirectory
boundaries as required by the spec:

  modules/auth/
    controllers/   auth.controller.ts   — HTTP handlers, rate-limiter state
    services/      token.service.ts     — JWT signing/verification
                   account-status.service.ts
                   metrics.service.ts
                   session-cleanup.service.ts
    repositories/  identity.repository.ts  — in-memory user store
                   session.repository.ts   — session + refresh-token index
    validators/    password.validator.ts
    routes/        index.ts             — thin Express router, maps URLs to handlers
    tests/                              — renamed from __tests__
    types/         index.ts             — AccountStatus, AuthIdentity

  shared/
    errors/        index.ts             — authErrorStatus, normalizeAuthError
    logger/        index.ts             — structured auth logger
    middleware/    auth-context.ts      — JWT verification middleware
                   response-helpers.ts  — unauthorized / forbidden helpers
                   role-guard.ts        — permission enforcement
    types/         role-policies.ts     — role → permission mapping
