+++ apps/api/src/modules/auth/services/session-cleanup.service.ts
@@ -1,0 +1,6 @@
+import { sessionStore } from "../repositories/session.repository.js";
+
+export function runSessionCleanup(): { purged: true; ranAt: string } {
+  sessionStore.purgeExpired();
+  return { purged: true, ranAt: new Date().toISOString() };
+}
