+++ apps/api/src/modules/auth/types/index.ts
@@ -1,0 +1,13 @@
+import type { UserRole } from "@lumen/types";
+
+export type AccountStatus = "pending" | "active" | "suspended" | "locked";
+
+export type AuthIdentity = {
+  userId: string;
+  clinicId: string;
+  email: string;
+  passwordHash: string;
+  role: UserRole;
+  status: AccountStatus;
+  createdAt: string;
+};
