+++ apps/api/src/shared/middleware/response-helpers.ts
